(feat): custom hook support for api data details
