+++ src/components/Details.jsx
@@ -1,28 +1,7 @@
-import { useEffect, useState } from "react";
+import useApiData from "../hooks/useApiData";
 
 const Details = (props) => {
-  const [repoDetails, setRepoDetails] = useState({
-    id: "",
-    name: "not found",
-    star: "not found",
-    language: "not found",
-    link: "not found",
-  });
-  useEffect(() => {
-    const addRepoDetails = async () => {
-      const url = props.repoLink;
-      const data = await fetch(url);
-      const newData = await data.json();
-      setRepoDetails({
-        id: newData.id,
-        name: newData.name,
-        language: newData.language,
-        link: newData.html_url,
-        star: newData.stargazers_count,
-      });
-    };
-    addRepoDetails();
-  }, []);
+  const repoDetails = useApiData(props.repoLink);
   return (
     <div className="item-center self-center">
       <a
@@ -61,7 +40,7 @@ const Details = (props) => {
           >
             <path d="M1 2.75C1 1.784 1.784 1 2.75 1h10.5c.966 0 1.75.784 1.75 1.75v7.5A1.75 1.75 0 0 1 13.25 12H9.06l-2.573 2.573A1.458 1.458 0 0 1 4 13.543V12H2.75A1.75 1.75 0 0 1 1 10.25Zm1.75-.25a.25.25 0 0 0-.25.25v7.5c0 .138.112.25.25.25h2a.75.75 0 0 1 .75.75v2.19l2.72-2.72a.749.749 0 0 1 .53-.22h4.5a.25.25 0 0 0 .25-.25v-7.5a.25.25 0 0 0-.25-.25Z"></path>
           </svg>
-          <p className="">16</p>
+          <p className="">{props.comments}</p>
         </div>
       </a>
     </div>
